Update HorizontalNav component to allow accessReview per-page.
Each page in HorizontalNav can now have an optional 'accessReview' property containing
SelfSubjectAccessReview resourceAttributes. If an accessReview property exists on a page, then the
associated page and tab will only be rendered if an SSAR with those resourceAttributes has
status.allowed of true. This will not affect pages that do not define an accessReview property.